File Structure: Rename sequencer --> video_editing
I did not want to make this change now but the reasoning is, it is better to make the changes now rather than later.

Currently URL look like: https://docs.blender.org/manual/en/dev/sequencer/sequencer/index.html

Which are a bit strange.

This change will allow us to expand this chapter more in the future.