Driver Workflow Examples: updated last shape key driver example.
The previous version of this example was bogus and had unclear goals.
There were instructions in the text and images to make shape keys relative to another key than the basis,
but these didn't match the video.
The example was refactored to focus only on the driver setup to chain shape keys.
The previous example also illustrated how to combine inputs from multiple variables,
but that was considered secondary.